[FIX] Blocking the check-enough-worker channel bug
The check-enough-worker channel will not be processed after number of
workers is enough. If we continuously register more workers, it will be
blocked by unbuffered channel, the check-enough-worker channel.

I use busy waiting method, check the number of workers and the number of
desired workers until the workers is enough.